domainとinfrastructureを作成
