+++ backend/domain/model/diary.go
@@ -1,0 +1,7 @@
+package model
+
+type Diary struct {
+	ID          int
+	Title       string
+	Description string
+}
